requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
python-dotenv
pydantic-settings
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
pypdf
argon2-cffi
pytest
httpx
pytest-asyncio
fastapi-mail
supabase
reportlab
qrcode
pillow
